fix/ci

+++ frontend/app/src/App.tsx
@@ -199,7 +199,7 @@ const App: FC = () => {
       ]);
 
       // Read the result
-      const fileData = await ffmpeg.readFile("output.mp4"); // Uint8Array (FileData)
+      const fileData = (await ffmpeg.readFile("output.mp4")) as Uint8Array;
 
       // Normalize to a plain Uint8Array so Blob typing works with strict TS
       const outputData: Uint8Array =
